+++ common/Utils.lua
@@ -1,19 +1,15 @@
 local addonName, EXT = ...
 
-local L = EXT.localization
+local L = EXT.Localization
 
 local Utils = {}
-
------------------------
---- Helper Funtions ---
------------------------
 
 ---------------------
 --- Main Funtions ---
 ---------------------
 
 function Utils:PrintDebug(msg)
-    if EXT.data.options["debug-mode"] then
+    if EXT.options.other["debug-mode"] then
 		DEFAULT_CHAT_FRAME:AddMessage(ORANGE_FONT_COLOR:WrapTextInColorCode(addonName .. " (Debug): ")  .. msg)
 	end
 end
@@ -23,13 +19,44 @@ function Utils:PrintMessage(msg)
 end
 
 function Utils:InitializeDatabase()
-    -- Options
-    if (not Expositum_Options) then
-        Expositum_Options = {}
+    if (not Expositum_Options_v2) then
+        Expositum_Options_v2 = {
+			["general"] = {
+				["minimap-button"] = {
+					["hide"] = false
+				}
+			},
+			["tooltip"] = {},
+			["other"] = {}
+		}
     end
 
-    EXT.data = {}
-    EXT.data.options = Expositum_Options
+    EXT.options = {}
+	EXT.options.general = Expositum_Options_v2["general"]
+	EXT.options.tooltip = Expositum_Options_v2["tooltip"]
+	EXT.options.other = Expositum_Options_v2["other"]
 end
 
-EXT.utils = Utils
+function Utils:InitializeMinimapButton()
+    local LDB = LibStub("LibDataBroker-1.1"):NewDataObject("Expositum", {
+        type     = "launcher",
+        text     = "Expositum",
+        icon     = EXT.MEDIA_PATH .. "icon-round.blp",
+        OnClick  = function(self, button)
+            if button == "RightButton" then
+                Settings.OpenToCategory(EXT.MAIN_CATEGORY_ID)
+            end
+        end,
+        OnTooltipShow = function(tooltip)
+			GameTooltip_SetTitle(tooltip, addonName)
+			GameTooltip_AddNormalLine(tooltip, EXT.ADDON_VERSION .. " (" .. EXT.ADDON_BUILD_DATE .. ")")
+			GameTooltip_AddBlankLineToTooltip(tooltip)
+			GameTooltip_AddHighlightLine(tooltip, L["minimap-button.tooltip"])
+        end,
+    })
+
+    self.minimapButton = LibStub("LibDBIcon-1.0")
+    self.minimapButton:Register("Expositum", LDB, EXT.options.general["minimap-button"])
+end
+
+EXT.Utils = Utils